refactor: Better gitlab stub (#1088)
* feat: Starting with a GitLab stub based on http4s

This commit only introduces the concept and migrates one test for
demonstration purposes.

* feat: Migrating `ProjectsResourcesSpec`

* refactor: Improve usage and design of GitLabApiStub syntax

More "low-level" operations are collected as functions in
`GitLabStateUpdates` and `GitLabStateQueries`, respectively.
Convenience syntax for writing tests are based on those and provided
in `GitLabStubIoSyntax`.

* feat: Migrate more tests to new GitLabStub

* feat: Fix adding commits, migrate CommitHistoryChangesSpec

* feat: Migrating CommitSyncFlowsSpec

* feat: Migrate EventFlowsSpec

* feat: Migrate EventsProcessingStatusSpec

* feat: Migrate FastTractEventRouteSpec

* chore: Add missing license headers

* feat: Migrate remaining specs

* chore: Deactivate previous GitLabStub

* refactor: More robust startup

* Fix EventsResourceSpec

* refactor: test trait GitLabStubSupport inherits the syntax

* refactor: remove now obsolete code

* refactor: pass the accessToken explicitely

Sometimes there has to be differentiated between different users

* refactor: Renamings

- AcceptanceSpec is now the base for all tests (was BaseSpec)
- rename GraphServices -> ApplicationServices
- pulled out DatasetsEncoders

* fix: Try fixing generators in tests

The generators used some random implicit value imported via
`.testentities._` or `EntitiesGenerators._`. But it requires to use
the same urls as used by the backend services. The gitlab-url must be
used from the config file in the acceptance tests.

* Refactor: move `findAllpersons` into `RenkuProject` companion