fix: deduplicate project created dates after transformation (#1236)
* Deduplicate project created dates after transformation
* Be aware of multiple dateCreated values on transformation

Refactoring deduplication of person to deterministically delete the
same values to ensure correctness in corner cases like concurrent
execution.

When loading a project during transformation, the dateCreated property
could exist multiple times. This is an edge case when not
deduplicating the dateCreated correctly. In this case, transformation
continues and will remove obsolete dateCreated values as a
post-processing step.