chore: Replace compatibility-matrix with a simpler config (#1268)
The compatibility-matrix was used to specify which cli-version was
compatible with a schema-version. Based on this a decision whether to
trigger re-provisioning was derived. Now the config only contains the
expected CLI version, the expected schema version and whether to kick
off re-provisioning if, and only if, cli or schema version is
different from the last run.