chore: Map dateModified to dateCreated
When CLI passes a dateModified, it copies dateCreated from the first
dataset or plan. We keep dateCreated on each entity as the date the
specific entity was created and so map the dateModified to our
dateCreated. To get the dateCreated of the original dataset/plan, we
can follow the `topMostSameAs` edge.