fix: decode child-plans of composites with same decoder as parent
Adds a scalatest matcher specific for jsonld decoding that prints the
raw jsonld in case of error. It prints it in one line, so we can copy
it out easily and inspect it.

The "lenient typed" decoder for composite plans was broken in that
while it was lenient to the types of a parent plan, it used the strict
decoders for its children. This lead to skipping child plans that had
additional types added.