feat: Don't require authenticated user for event/statuts endpoint (#1451)
* Introduce MaybeAuthUser, refactoring accordingly

This is to better distinguish the case "no token provided" vs "token
invalid" when reading code. The `…ifNeeded` auth middleware is
fulfilling the route always, either with a `None` or `Some` based on
the existence of a token. That means the routes cannot be combined via
semigroupK anymore, since there is no "fall through" anymore. This
also applies to the mandatory auth middleware. That means both ways
cannot be combined via "the usual way". To overcome this, the
`MaybeUser` provides some convenient methods to support cases where a
user is mandatory and optional, so routes can be combined.