feat: new datasets query (#1398)
Adds a query for datasets that uses the separate graph in order to speed up search. Using this requires to return a different id for a dataset that maps to the "same-as" property. Since this changes the data structures of the public api, this feature is disabled by default. It can be enabled by adding a request header `Renku-Kg-QueryNew` to the search request.