Try fixing acceptance tests
The expected dateModified is now a bit difficult to obtain. The cli
payload doesn't contain the dateModified, but it is derived from all
dates inside the payload. But when setting up the tests, we generate
some dateModified and put it into the gitlabs response. It could be
newer and then the merge process would take the newest of both.