feat: Return datasets dateModified in entity search (#1612)
* Return datasets dateModified in entity search
* Return dateModified from dataset details

Moves the value returned as dateCreated to be dateModified. And the
topmost-derived-from dataset is used to retrieve the created or
published date.

* Add redundant dateCreated/datePublished

For projects, this should replace the generic named "date" property.
For datasets, it provides additional more exact properties.

* Try fixing acceptance tests

The expected dateModified is now a bit difficult to obtain. The cli
payload doesn't contain the dateModified, but it is derived from all
dates inside the payload. But when setting up the tests, we generate
some dateModified and put it into the gitlabs response. It could be
newer and then the merge process would take the newest of both.