chore: Remove allowed projects context (#1822)
* Remove usage of authContext#allowedProjects

It's only used in BaseDetailsFinder, and it can be replaced by the
snippet going into the ProjectAuth graph. The unit tests of the
other *Finders use it and are sometimes not setup correctly. For
example, they create public projects but restrict visibilty by setting
random "allowed projects". With this change the tests need to be setup
more carefully so that the data really represents the tested cases.