Moving things around
The idea is to have more lighter/smaller modules in the "utils"
directory, each aiming at one (or not too many) purposes. The
"commonUtils" sevres as the base dependency.

The dependency tiny-types->triples-store-client has been removed. The
functionality of both is currently in renku-model-tiny-types as a
temporary solution.

Another change in tiny types: The factory inherits from many traits,
one is `From` to lift a "normal" value into a tiny type. In order to
provide these implementations from other modules, I used it now as a
type class. This way we can provide more functionality requiring a
`From` - like jsonld decoders.

It's very rough still, serving now as a starting point to have a
better feeling where and how to go forward and if that is the desired
way :).